requirements.txt
attrs==19.3.0
Automat==20.2.0
certifi==2021.10.8
cffi==1.15.0
colormath==3.0.0
constantly==15.1.0
cycler==0.10.0
Cython==0.29.24
future==0.18.2
glcontext==2.3.4
hyperlink==21.0.0
idna==3.3
incremental==21.3.0
kiwisolver==1.3.2
kociemba==1.2.1
matplotlib==3.4.3
moderngl==5.6.4
moderngl-window==2.4.0
multipledispatch==0.6.0
networkx==2.6.3
numpy==1.21.3
opencv-python==4.5.4.58
Panda3D==1.10.10
panda3d-gltf==0.13
panda3d-simplepbr==0.9
pandas==1.3.4
Pillow==8.4.0
projector==0.1.1
pycparser==2.20
pygame==2.0.2
pyglet==1.5.21
PyHamcrest==2.0.2
pyobjc-core==7.3
pyobjc-framework-Cocoa==7.3
PyOpenGL==3.1.5
pyparsing==3.0.1
pyperclip==1.8.2
pyproj==3.2.1
PyQt5==5.15.5
PyQt5-Qt5==5.15.2
PyQt5-sip==12.9.0
PyQt5-stubs==5.15.2.0
pyquaternion==0.9.9
pyrr==0.10.3
python-dateutil==2.8.2
pytz==2021.3
PyYAML==6.0
rubik-cube==0.0.1
screeninfo==0.7
six==1.16.0
turtle==0.0.1
ursina==4.0.0
zope.interface==5.4.0